feat: db를 postgresql로 변경 및 postgresql에 맞는 쿼리문으로 수정

+++ src/timeline/participation.entity.ts
@@ -13,7 +13,7 @@ export class Participation {
   @PrimaryGeneratedColumn()
   participationId: number
 
-  @Column({ type: 'int', default: 0, nullable: false })
+  @Column({ type: 'float', default: 0, nullable: false })
   playHour: number
 
   @ManyToOne(() => Streamer, (streamer) => streamer.participations, {
